now every js file contains only one class
