+++ index.html
@@ -5,10 +5,8 @@
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-  <!--link rel="stylesheet" type="text/css" href="ReasonRoot.css" media="screen"/-->
-
-
-  <!--This scripts importation was substituted by the entry propriety on the webpack.config.js file.>
+  <!--This scripts and links importation was substituted by the entry propriety on the webpack.config.js file.>
+  <!link rel="stylesheet" type="text/css" href="ReasonRoot.css" media="screen"/>
   <script src="js/src/rrDisplay.js"></script>
   <script src="js/src/Claim.js"></script>
   <script src="js/src/score.js"></script>
